requirements.txt
discord.py==2.3.2
pandas==2.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
